add new function

+++ fourinsight/engineroom/utils/_core.py
@@ -3,6 +3,7 @@ from abc import abstractmethod
 from collections.abc import MutableMapping
 from io import BytesIO, TextIOWrapper
 from pathlib import Path
+import urllib.parse
 
 import pandas as pd
 from azure.core.exceptions import ResourceNotFoundError
@@ -549,7 +550,7 @@ class ResultCollector:
         if index_drop:
             self.delete_rows(index_drop)
 
-def load_past_er_results(app_id, session, path=None, download_all=False, output_folder="output"):
+def load_previous_engineroom_results(app_id, session, path=None, download_all=False, output_folder="output"):
     """
     Load past EngineRoom results from a specified application and path and 
     store locally in the same output folder
@@ -571,25 +572,38 @@ def load_past_er_results(app_id, session, path=None, download_all=False, output_
         Name of the EngineRoom output folder. Defaults to "output".
 
     """
-    #ensure output folder exist
-    output_folder = Path(f"./{output_folder}")
-    output_folder.mkdir(parents=True, exist_ok=True)
+    output_folder = Path(output_folder)
+
+    existing_oputput_files = session.get(f"https://api.4insight.io/v1.0/Applications/{app_id}/results")
+    existing_oputput_files.raise_for_status()
+    
+    if not existing_oputput_files.json():
+        raise ValueError(f"No results found for application ID {app_id}.")
+
+    available_file_names = [file["fileName"] for file in existing_oputput_files.json()]
+    navigable_file_names = [file["navigableFileName"] for file in existing_oputput_files.json()]
 
     if download_all == True:
-        existing_oputput_files = session.get(f"https://api.4insight.io/v1.0/Applications/{app_id}/results")
-        existing_oputput_files.raise_for_status()
-        for file in existing_oputput_files.json():
-            path = file["fileName"]
+        for path in available_file_names:
             _path = output_folder / path
-            output_url = f"https://api.4insight.io/v1.0/Applications/{app_id}/results/{path}/download"
+            _path.parent.mkdir(parents=True, exist_ok=True)
+
+            navigable_file_name = urllib.parse.quote(navigable_file_names[available_file_names.index(path)])
+
+            output_url = f"https://api.4insight.io/v1.0/Applications/{app_id}/results/{navigable_file_name}/download"
             previous_results = session.get(output_url)
             previous_results.raise_for_status()
 
             with open(_path, "wb") as download_file:
                 download_file.write(previous_results.content)
     else:
+        if path not in available_file_names:
+            raise ValueError(f"{path} not found in application {app_id} results.")
         _path = output_folder / path
-        output_url = f"https://api.4insight.io/v1.0/Applications/{app_id}/results/{path}/download"
+        _path.parent.mkdir(parents=True, exist_ok=True)
+        
+        navigable_file_name = urllib.parse.quote(navigable_file_names[available_file_names.index(path)])
+        output_url = f"https://api.4insight.io/v1.0/Applications/{app_id}/results/{navigable_file_name}/download"
         previous_results = session.get(output_url)
         previous_results.raise_for_status()
